logging: if you don't set root logger's level, handler levels will be ignored.
see this SO:
https://stackoverflow.com/questions/17668633/what-is-the-point-of-setlevel-in-a-python-logging-handler

Essentially:

- the logging.basicConfig() call sets properties for the root logger.
- if you don't supply the 'level' property (which sets the log level),
  it will default to WARNING. So the root logger will have a log level of WARNING.
- for individual handlers, when you call .setLevel(<level>) on them,
  it will look at the root handler's level, and if it is LESS than that, it will
  ignore it, and assign as a level whatever the root logger's level is.

So as an example:
Suppose you don't set set root logger's level.
It will default to WARNING.
Now suppose you try to set the level of your StreamHandler (what prints to console)
 as logging.DEBUG.
Since logging.DEBUG < logging.WARNING it will ignore it, and the StreamHandler
will get assigned a log level of logging.WARNING (the root logger's level.)

The solution here is to ALWAYS set the log level of the root logger,
and set it as low as possible. Because, in our case, we are relying on our
individual handlers (FileHandler and StreamHandler) so the root logger level
is kind of worthless to us.