Introduced bug to copy_file; revamp createPath
- In commit ed87f5482dda05c4bf2730aff8a8f9526055e4a0 I revamped
"copy_file" function. Recall that "copy_file" takes args
for "src" (source file to copy) and "dest" (dest to copy it to),
and that dest can be either a directory OR a file
(i.e. that you either want to copy it into some directory,
or essentially copy and rename it to something.)

I redid the logic to handle the situation where the dest to
copy to doesn't exist and doesn't have a file extension;
in such a case, there's ambiguitity if dest is meant to be
a directory, or a file without an extension.

I broke things as a result...

Rather than explaining the logic here, just submitting the commit
because the logic takes a long time to describe.

Other things:

- adding function file_ext to return file extension, as I'm
always forgetting how to do it, or having to add comments
explaining what I'm doing since it's not obvious