Bug: Fix removal of trailing / on void tags (1315c05 error)
- Commit 1315c05 gave an option in write_soup_to_file to remove
trailing / in void tags such as <br> and <img> (as BeautifulSoup
adds these in during prettify, and it causes w3 validation errors).

- In that commit, did it naively, just looking for all occurrances
of /> and replacing with >. However, this is a regression: /> is
a valid way to close non-void tags, and so removing this can mangle
HTML. (Example: when adding in icons with <svg>, there is a <path>
tag within that ends in />, and removing that ruins the icon.
See an example here:
https://vocab.rutools.xyz/vocab/telling_time.html
(for the icon on the tooltip)

- This fixes that issue by only doing the replace at end of
void tags.