test_bs.py: put tests in functions and call them all
- this is a messy file that i built over time. there's no functions
and the tests are just executed one after another
- put them all in functions with test_ prefix
- have a main function where it looks for all functions prefixed
with test_ and calls them all

so ultimately, if you call python test_bs.py, it executes all
the functions just like it currently does.

(the point of this is to ultimately streamline this into a testing
framework. and when i do that, i want it to be in this format
so that the commit change is minimal.)